# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SSSD-APP API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://b-shop-api-5ntn.onrender.com/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get list of orders",
                "description": "Returns list of orders",
                "operationId": "325fd3d55bbe9bb39645269928fa4983",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Order"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Create a new order",
                "description": "Create a new order",
                "operationId": "445bab4f7dc9ea6fcd54f286c0209478",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Order"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/orders/{id}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get specified order",
                "description": "Returns specified order",
                "operationId": "f088d34a2737782215f435f854e31de5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the order",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Orders"
                ],
                "summary": "Update specified order",
                "description": "Update specified order",
                "operationId": "54c2858108887d7a1252bcc4325158c2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the order",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Order"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Orders"
                ],
                "summary": "Delete specified order",
                "description": "Delete specified order",
                "operationId": "deb80b1c70721ac0812c9574580984c1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the order",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/payments": {
            "get": {
                "tags": [
                    "Payments"
                ],
                "summary": "Get list of payments",
                "description": "Returns list of payments",
                "operationId": "65567ad73dd4b0cc01c4856a6bb7e600",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Payment"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Payments"
                ],
                "summary": "Create a new payment",
                "description": "Create a new payment",
                "operationId": "3cb7bc2c2ad5498b4ad9521811750cc1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Payment"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Payment"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/payments/{id}": {
            "get": {
                "tags": [
                    "Payments"
                ],
                "summary": "Get specified payment",
                "description": "Returns specified payment",
                "operationId": "a4207c19fecbebc99932a8348b19adf1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the payment",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Payment"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Payments"
                ],
                "summary": "Update specified payment",
                "description": "Update specified payment",
                "operationId": "7f3042f886092282c90cf678fa8b0d78",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the payment",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Payment"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Payment"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Payments"
                ],
                "summary": "Delete specified payment",
                "description": "Delete specified payment",
                "operationId": "6ee6ad54e3a047f39a06ffcb4cfc6b7a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the payment",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Order": {
                "title": "Order",
                "description": "Order model",
                "properties": {
                    "id": {
                        "description": "Unique identifier for the order",
                        "type": "integer",
                        "example": "1"
                    },
                    "user_id": {
                        "description": "User ID associated with the order",
                        "type": "integer",
                        "example": "1"
                    },
                    "product_id": {
                        "description": "Product ID associated with the order",
                        "type": "integer",
                        "example": "1"
                    },
                    "quantity": {
                        "description": "Quantity of the product in the order",
                        "type": "integer",
                        "example": "2"
                    },
                    "order_status": {
                        "description": "Status of the order",
                        "type": "string",
                        "example": "pending"
                    },
                    "tracking_status": {
                        "description": "Status of the order tracking",
                        "type": "string",
                        "example": "shipped"
                    },
                    "created_at": {
                        "description": "Date and time when the order was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-05-12 07:00:00"
                    },
                    "updated_at": {
                        "description": "Date and time when the order was last updated",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-05-12 07:30:00"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "payment": {
                        "$ref": "#/components/schemas/Payment"
                    }
                },
                "type": "object"
            },
            "Payment": {
                "title": "Payment",
                "description": "Payment model",
                "properties": {
                    "id": {
                        "description": "Unique identifier for the payment",
                        "type": "integer",
                        "example": "1"
                    },
                    "order_id": {
                        "description": "Order ID associated with the payment",
                        "type": "integer",
                        "example": "1"
                    },
                    "payment_method": {
                        "description": "Payment method used",
                        "type": "string",
                        "example": "credit_card"
                    },
                    "payment_status": {
                        "description": "Status of the payment",
                        "type": "string",
                        "example": "completed"
                    },
                    "payment_amount": {
                        "description": "Amount of the payment",
                        "type": "number",
                        "format": "float",
                        "example": "100.00"
                    },
                    "buyer_id": {
                        "description": "ID of the buyer",
                        "type": "integer",
                        "example": "1"
                    },
                    "seller_id": {
                        "description": "ID of the seller",
                        "type": "integer",
                        "example": "2"
                    },
                    "created_at": {
                        "description": "Date and time when the payment was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-05-12 06:00:00"
                    },
                    "updated_at": {
                        "description": "Date and time when the payment was last updated",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-05-12 06:30:00"
                    },
                    "order": {
                        "$ref": "#/components/schemas/Order"
                    }
                },
                "type": "object"
            },
...
            "User": {
                "title": "User",
                "description": "User model",
                "properties": {
                    "id": {
                        "description": "The unique identifier for the user",
                        "type": "integer",
                        "example": "1"
                    },
                    "full_name": {
                        "description": "The full name of the user",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "username": {
                        "description": "The username of the user",
                        "type": "string",
                        "example": "johndoe"
                    },
                    "email": {
                        "description": "The email address of the user",
                        "type": "string",
                        "format": "email",
                        "example": "johndoe@example.com"
                    },
                    "password": {
                        "description": "The password of the user (hashed)",
                        "type": "string",
                        "format": "password",
                        "example": "$2y$10$N0wbN10w.0.yBhEWwH2fveCAeAxFxuIQJ4O7QHBTQm7v9eN6LpfSi"
                    },
                    "phone_number": {
                        "description": "The phone number of the user",
                        "type": "string",
                        "example": "+1234567890"
                    },
                    "email_verified_at": {
                        "description": "The timestamp of email verification",
                        "type": "string",
                        "format": "date-time",
                        "example": "2022-03-28T10:15:00Z"
                    },
                    "created_at": {
                        "description": "The timestamp of when the user was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2022-03-28T10:15:00Z"
                    },
                    "updated_at": {
                        "description": "The timestamp of when the user was last updated",
                        "type": "string",
                        "format": "date-time",
                        "example": "2022-03-28T10:15:00Z"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}